Wizard Spellbook QoL (#76872)
Second verse, same as the first

Alphabetizes the wizard spellbook, adds a search bar.

![image](https://github.com/tgstation/tgstation/assets/51863163/5c94aed6-8010-435e-8c37-a7d24d6b9ab6)

![image](https://github.com/tgstation/tgstation/assets/51863163/5bd3b004-a4ad-4904-b76c-958e473f1e25)

Makes it easier to buy stuff

:cl: Melbert
qol: Wizard Spellbook is now alphabetized
qol: Wizard Spellbook now has a search bar!
/:cl:
